Create `integration-tests/internal/builtin-precompiles`
